implemented account selector as a standalone plugin moved account selector to a different package
